+++ mockify/mock/_mock.py
@@ -1,7 +1,7 @@
 # ---------------------------------------------------------------------------
 # mockify/mock/_mock.py
 #
-# Copyright (C) 2019 - 2020 Maciej Wiatrzyk <maciej.wiatrzyk@gmail.com>
+# Copyright (C) 2019 - 2021 Maciej Wiatrzyk <maciej.wiatrzyk@gmail.com>
 #
 # This file is part of Mockify library and is released under the terms of the
 # MIT license: http://opensource.org/licenses/mit-license.php.
@@ -9,6 +9,8 @@
 # See LICENSE for details.
 # ---------------------------------------------------------------------------
 from typing import Callable, Dict
+
+from mockify.interface import MockAttr
 
 from .. import _utils
 from ..core import BaseMock, Call, MockInfo
@@ -72,6 +74,13 @@ class Mock(BaseMock):
     @property
     def _info(self):
         return MockInfo(self)
+
+    def __m_get_method__(self, attr: MockAttr) -> 'BaseMock':
+        path = attr.path()
+        mock = getattr(self, next(path))
+        for name in path:
+            mock = getattr(mock, name)
+        return mock
 
     def __m_children__(self):
         for obj in self.__dict__.values():